requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
astroid==2.4.2
billiard==3.6.3.0
celery==4.4.7
Django==3.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-rest-framework==0.1.0
django-timezone-field==4.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
isort==4.3.21
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.5.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
